Added `event.context` for response to 'callbackURL' response to make the backend aware of the `record_id and class` of the original file.
Do not initialize redundant local variables for `callbackURL` and `versions` for single usage - just referencing them from the `event` object is enough.

`targetStorage` is sent not in the `original` file's context, but as an `event`'s key.

Changed the 'callbackURL' in the sample event.json file to a `localtunnel.me` URL which is working crossplatform.

Added to README.md the instruction to install `aws-sam-local` package.